Remove some now-obsolete casts

+++ src/core/geometry/PixelDetectorModel.cpp
@@ -49,7 +49,7 @@ bool PixelDetectorModel::isWithinSensor(const ROOT::Math::XYZPoint& local_pos) c
 bool PixelDetectorModel::isWithinImplant(const ROOT::Math::XYZPoint& local_pos) const {
 
     auto [xpixel, ypixel] = getPixelIndex(local_pos);
-    auto inPixelPos = local_pos - getPixelCenter(static_cast<unsigned int>(xpixel), static_cast<unsigned int>(ypixel));
+    auto inPixelPos = local_pos - getPixelCenter(xpixel, ypixel);
 
     return (std::fabs(inPixelPos.x()) <= std::fabs(getImplantSize().x() / 2) &&
             std::fabs(inPixelPos.y()) <= std::fabs(getImplantSize().y() / 2));
